Rename scripts to stub_uploader
